fix: click on plot doesn't reset zoom + zoom horizontal only
Since the plotly plots are not fully react (the state is not kept),
using react state for showing text on click (plot by day) re-renders
the plot, which loose all current zoom/pan, etc.

The fix is to use plain JS to update the text. This is not really
clean, but the only way around.

Also, fix the y axis on plots, so zoom only goes horizontal,
and disable dragmode on weekdays, since it would make no sense to
zoom here.